removeUnusedDeclarationsTest
